+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+
+## [0.1.1] - 2023-11-08
+
+**Implemented enhancements:**
+- Added cluster name in the report details
+
+### Fixed
+- The alive message is sent directly to the queue
+
+
 ### First Release
 
 v0.1.0: [https://github.com/seriohub/velero-watchdog-pro/releases/tag/v0.1.0](https://github.com/seriohub/velero-watchdog-pro/releases/tag/v0.1.0)